+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Change Log
 This project follows [semver 2.0.0][1] and the recommendations
 of [keepachangelog.com][2].
 
-0.0.2 (unreleased)
+0.0.3
 -----
 
 ### Changed
@@ -16,6 +16,12 @@ of [keepachangelog.com][2].
 - Configuration
   - Added `:no_paren_methods` omit argument parentheses
 
+0.0.2
+-----
+
+### Changed
+- Normalize block arguments.  See [ruby_parser PR 189][3] (Ryan Davis)
+
 0.0.1
 -----
 
@@ -23,3 +29,4 @@ Initial version.  Just claiming the gem name.
 
 [1]: http://semver.org/spec/v2.0.0.html
 [2]: http://keepachangelog.com/
+[3]: https://github.com/seattlerb/ruby_parser/pull/189